+++ service_area/provider/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class ProviderConfig(AppConfig):
+
+    name = 'service_area.provider'